@@ -1,50 +1,50 @@
 <?php
 
-namespace TFox\MpdfPortBundle\Service;
+namespace Akyos\MpdfBundle\Service;
 
+use ReflectionClass;
 use Symfony\Component\HttpFoundation\Response;
 
 class MpdfService {
 	
-    private $addDefaultConstructorArgs = true;
-	
+    private bool $addDefaultConstructorArgs = true;
+
     /**
      * Get an instance of mPDF class
      * @param array $constructorArgs arguments for mPDF constror
      * @return \mPDF
+     * @throws \ReflectionException
      */
-    public function getMpdf($constructorArgs = array()) 
+    public function getMpdf(array $constructorArgs = []): \mPDF
     {
         $allConstructorArgs = $constructorArgs;
 	    if($this->getAddDefaultConstructorArgs()) {
-	        $allConstructorArgs = array(array_merge(array('utf-8', 'A4'), $allConstructorArgs));
+	        $allConstructorArgs = [array_merge(['utf-8', 'A4'], $allConstructorArgs)];
 	    }		
 
-        $reflection = new \ReflectionClass('Mpdf\Mpdf');
-        $mpdf = $reflection->newInstanceArgs($allConstructorArgs);
+        $reflection = new ReflectionClass('Mpdf\Mpdf');
 
-        return $mpdf;
+        return $reflection->newInstanceArgs($allConstructorArgs);
     }	     
 	
     /**
      * Returns a string which content is a PDF document
      * @param string $html Html content
      * @param array $argOptions Options for the constructor.
-     *
      * @return PDF file.
      */
-    public function generatePdf($html, array $argOptions = array())
+    public function generatePdf(string $html, array $argOptions = []): PDF
     {
         //Calculate arguments
-        $defaultOptions = array(
-            'constructorArgs'      => array(),
+        $defaultOptions = [
+            'constructorArgs'      => [],
             'writeHtmlMode'        => null,
             'writeHtmlInitialise'  => null,
             'writeHtmlClose'       => null,
             'outputFilename'       => '',
             'outputDest'           => 'S',
             'mpdf'                 => null
-        );                
+        ];
 	
         $options = array_merge($defaultOptions, $argOptions);
         extract($options);
@@ -54,27 +54,26 @@ class MpdfService {
         }
 
         //Add arguments to AddHtml function
-        $writeHtmlArgs = array($writeHtmlMode, $writeHtmlInitialise, $writeHtmlClose);
+        $writeHtmlArgs = [$writeHtmlMode, $writeHtmlInitialise, $writeHtmlClose];
         $writeHtmlArgs = array_filter($writeHtmlArgs, function($x) { 
             return !is_null($x); 
         });
 
         $writeHtmlArgs['html'] = $html;
 
-        @call_user_func_array(array($mpdf, 'WriteHTML'), $writeHtmlArgs);
+        @call_user_func_array([$mpdf, 'WriteHTML'], $writeHtmlArgs);
 
         //Add arguments to Output function
-        $content = $mpdf->Output($outputFilename, $outputDest);
-	
-        return $content;
+        return $mpdf->Output($outputFilename, $outputDest);
     }
-	
+
     /**
      * Generates an instance of Response class with PDF document
      * @param string $html
      * @param array $argOptions
+     * @return \Symfony\Component\HttpFoundation\Response
      */
-    public function generatePdfResponse($html, array $argOptions = array())
+    public function generatePdfResponse(string $html, array $argOptions = []): Response
     {
         $response = new Response();		
         $response->headers->set('Content-Type', 'application/pdf');
@@ -84,12 +83,13 @@ class MpdfService {
 
         return $response;
     }
-	
-   /**
-    * Set the defaults argumnets to the constructor.
-    * @param $array $val
-    */
-    public function setAddDefaultConstructorArgs($val)
+
+    /**
+     * Set the defaults argumnets to the constructor.
+     * @param $val
+     * @return \Akyos\MpdfBundle\Service\MpdfService
+     */
+    public function setAddDefaultConstructorArgs($val): static
     {
         $this->addDefaultConstructorArgs = $val;
         
@@ -99,7 +99,7 @@ class MpdfService {
    /**
     * Get defaults arguntments.
     */
-    public function getAddDefaultConstructorArgs()
+    public function getAddDefaultConstructorArgs(): bool
     {
 	    return $this->addDefaultConstructorArgs;
     }
